requirements/base.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-access-policy==1.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
drf-standardized-errors==0.14.1
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==8.13.53
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
pyparsing==3.2.3
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.0
uritemplate==4.1.1
